+++ include/libqabstractanimation.zig
@@ -35,21 +35,21 @@ pub const qabstractanimation = struct {
     }
 
     /// ``` self: QtC.QAbstractAnimation, param1: qobjectdefs_enums.Call, param2: i32, param3: ?*anyopaque ```
-    pub fn Metacall(self: ?*anyopaque, param1: i64, param2: i32, param3: ?*anyopaque) i32 {
+    pub fn Metacall(self: ?*anyopaque, param1: i32, param2: i32, param3: ?*anyopaque) i32 {
         return qtc.QAbstractAnimation_Metacall(@ptrCast(self), @intCast(param1), @intCast(param2), param3);
     }
 
     /// Allows for overriding the related default method
     ///
     /// ``` self: QtC.QAbstractAnimation, slot: fn (self: QtC.QAbstractAnimation, param1: qobjectdefs_enums.Call, param2: i32, param3: ?*anyopaque) callconv(.c) i32 ```
-    pub fn OnMetacall(self: ?*anyopaque, slot: fn (?*anyopaque, i64, i32, ?*anyopaque) callconv(.c) i32) void {
+    pub fn OnMetacall(self: ?*anyopaque, slot: fn (?*anyopaque, i32, i32, ?*anyopaque) callconv(.c) i32) void {
         qtc.QAbstractAnimation_OnMetacall(@ptrCast(self), @as(isize, @bitCast(@intFromPtr(&slot))));
     }
 
     /// Base class method implementation
     ///
     /// ``` self: QtC.QAbstractAnimation, param1: qobjectdefs_enums.Call, param2: i32, param3: ?*anyopaque ```
-    pub fn QBaseMetacall(self: ?*anyopaque, param1: i64, param2: i32, param3: ?*anyopaque) i32 {
+    pub fn QBaseMetacall(self: ?*anyopaque, param1: i32, param2: i32, param3: ?*anyopaque) i32 {
         return qtc.QAbstractAnimation_QBaseMetacall(@ptrCast(self), @intCast(param1), @intCast(param2), param3);
     }
 
@@ -70,7 +70,7 @@ pub const qabstractanimation = struct {
     /// ``` self: QtC.QAbstractAnimation ```
     ///
     /// Returns: ``` qabstractanimation_enums.State ```
-    pub fn State(self: ?*anyopaque) i64 {
+    pub fn State(self: ?*anyopaque) i32 {
         return qtc.QAbstractAnimation_State(@ptrCast(self));
     }
 
@@ -86,14 +86,14 @@ pub const qabstractanimation = struct {
     /// ``` self: QtC.QAbstractAnimation ```
     ///
     /// Returns: ``` qabstractanimation_enums.Direction ```
-    pub fn Direction(self: ?*anyopaque) i64 {
+    pub fn Direction(self: ?*anyopaque) i32 {
         return qtc.QAbstractAnimation_Direction(@ptrCast(self));
     }
 
     /// [Qt documentation](https://doc.qt.io/qt-6/qabstractanimation.html#setDirection)
     ///
     /// ``` self: QtC.QAbstractAnimation, direction: qabstractanimation_enums.Direction ```
-    pub fn SetDirection(self: ?*anyopaque, direction: i64) void {
+    pub fn SetDirection(self: ?*anyopaque, direction: i32) void {
         qtc.QAbstractAnimation_SetDirection(@ptrCast(self), @intCast(direction));
     }
 
@@ -181,14 +181,14 @@ pub const qabstractanimation = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qabstractanimation.html#stateChanged)
     ///
     /// ``` self: QtC.QAbstractAnimation, newState: qabstractanimation_enums.State, oldState: qabstractanimation_enums.State ```
-    pub fn StateChanged(self: ?*anyopaque, newState: i64, oldState: i64) void {
+    pub fn StateChanged(self: ?*anyopaque, newState: i32, oldState: i32) void {
         qtc.QAbstractAnimation_StateChanged(@ptrCast(self), @intCast(newState), @intCast(oldState));
     }
 
     /// [Qt documentation](https://doc.qt.io/qt-6/qabstractanimation.html#stateChanged)
     ///
     /// ``` self: QtC.QAbstractAnimation, slot: fn (self: QtC.QAbstractAnimation, newState: qabstractanimation_enums.State, oldState: qabstractanimation_enums.State) callconv(.c) void ```
-    pub fn OnStateChanged(self: ?*anyopaque, slot: fn (?*anyopaque, i64, i64) callconv(.c) void) void {
+    pub fn OnStateChanged(self: ?*anyopaque, slot: fn (?*anyopaque, i32, i32) callconv(.c) void) void {
         qtc.QAbstractAnimation_Connect_StateChanged(@ptrCast(self), @as(isize, @bitCast(@intFromPtr(&slot))));
     }
 
@@ -209,14 +209,14 @@ pub const qabstractanimation = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qabstractanimation.html#directionChanged)
     ///
     /// ``` self: QtC.QAbstractAnimation, param1: qabstractanimation_enums.Direction ```
-    pub fn DirectionChanged(self: ?*anyopaque, param1: i64) void {
+    pub fn DirectionChanged(self: ?*anyopaque, param1: i32) void {
         qtc.QAbstractAnimation_DirectionChanged(@ptrCast(self), @intCast(param1));
     }
 
     /// [Qt documentation](https://doc.qt.io/qt-6/qabstractanimation.html#directionChanged)
     ///
     /// ``` self: QtC.QAbstractAnimation, slot: fn (self: QtC.QAbstractAnimation, param1: qabstractanimation_enums.Direction) callconv(.c) void ```
-    pub fn OnDirectionChanged(self: ?*anyopaque, slot: fn (?*anyopaque, i64) callconv(.c) void) void {
+    pub fn OnDirectionChanged(self: ?*anyopaque, slot: fn (?*anyopaque, i32) callconv(.c) void) void {
         qtc.QAbstractAnimation_Connect_DirectionChanged(@ptrCast(self), @as(isize, @bitCast(@intFromPtr(&slot))));
     }
 
@@ -315,7 +315,7 @@ pub const qabstractanimation = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qabstractanimation.html#updateState)
     ///
     /// ``` self: QtC.QAbstractAnimation, newState: qabstractanimation_enums.State, oldState: qabstractanimation_enums.State ```
-    pub fn UpdateState(self: ?*anyopaque, newState: i64, oldState: i64) void {
+    pub fn UpdateState(self: ?*anyopaque, newState: i32, oldState: i32) void {
         qtc.QAbstractAnimation_UpdateState(@ptrCast(self), @intCast(newState), @intCast(oldState));
     }
 
@@ -324,7 +324,7 @@ pub const qabstractanimation = struct {
     /// Allows for overriding the related default method
     ///
     /// ``` self: QtC.QAbstractAnimation, slot: fn (self: QtC.QAbstractAnimation, newState: qabstractanimation_enums.State, oldState: qabstractanimation_enums.State) callconv(.c) void ```
-    pub fn OnUpdateState(self: ?*anyopaque, slot: fn (?*anyopaque, i64, i64) callconv(.c) void) void {
+    pub fn OnUpdateState(self: ?*anyopaque, slot: fn (?*anyopaque, i32, i32) callconv(.c) void) void {
         qtc.QAbstractAnimation_OnUpdateState(@ptrCast(self), @as(isize, @bitCast(@intFromPtr(&slot))));
     }
 
@@ -333,14 +333,14 @@ pub const qabstractanimation = struct {
     /// Base class method implementation
     ///
     /// ``` self: QtC.QAbstractAnimation, newState: qabstractanimation_enums.State, oldState: qabstractanimation_enums.State ```
-    pub fn QBaseUpdateState(self: ?*anyopaque, newState: i64, oldState: i64) void {
+    pub fn QBaseUpdateState(self: ?*anyopaque, newState: i32, oldState: i32) void {
         qtc.QAbstractAnimation_QBaseUpdateState(@ptrCast(self), @intCast(newState), @intCast(oldState));
     }
 
     /// [Qt documentation](https://doc.qt.io/qt-6/qabstractanimation.html#updateDirection)
     ///
     /// ``` self: QtC.QAbstractAnimation, direction: qabstractanimation_enums.Direction ```
-    pub fn UpdateDirection(self: ?*anyopaque, direction: i64) void {
+    pub fn UpdateDirection(self: ?*anyopaque, direction: i32) void {
         qtc.QAbstractAnimation_UpdateDirection(@ptrCast(self), @intCast(direction));
     }
 
@@ -349,7 +349,7 @@ pub const qabstractanimation = struct {
     /// Allows for overriding the related default method
     ///
     /// ``` self: QtC.QAbstractAnimation, slot: fn (self: QtC.QAbstractAnimation, direction: qabstractanimation_enums.Direction) callconv(.c) void ```
-    pub fn OnUpdateDirection(self: ?*anyopaque, slot: fn (?*anyopaque, i64) callconv(.c) void) void {
+    pub fn OnUpdateDirection(self: ?*anyopaque, slot: fn (?*anyopaque, i32) callconv(.c) void) void {
         qtc.QAbstractAnimation_OnUpdateDirection(@ptrCast(self), @as(isize, @bitCast(@intFromPtr(&slot))));
     }
 
@@ -358,7 +358,7 @@ pub const qabstractanimation = struct {
     /// Base class method implementation
     ///
     /// ``` self: QtC.QAbstractAnimation, direction: qabstractanimation_enums.Direction ```
-    pub fn QBaseUpdateDirection(self: ?*anyopaque, direction: i64) void {
+    pub fn QBaseUpdateDirection(self: ?*anyopaque, direction: i32) void {
         qtc.QAbstractAnimation_QBaseUpdateDirection(@ptrCast(self), @intCast(direction));
     }
 
@@ -391,7 +391,7 @@ pub const qabstractanimation = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qabstractanimation.html#start)
     ///
     /// ``` self: QtC.QAbstractAnimation, policy: qabstractanimation_enums.DeletionPolicy ```
-    pub fn Start1(self: ?*anyopaque, policy: i64) void {
+    pub fn Start1(self: ?*anyopaque, policy: i32) void {
         qtc.QAbstractAnimation_Start1(@ptrCast(self), @intCast(policy));
     }
 
@@ -507,7 +507,7 @@ pub const qabstractanimation = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qobject.html#killTimer)
     ///
     /// ``` self: QtC.QAbstractAnimation, id: qnamespace_enums.TimerId ```
-    pub fn KillTimer2(self: ?*anyopaque, id: i64) void {
+    pub fn KillTimer2(self: ?*anyopaque, id: i32) void {
         qtc.QObject_KillTimer2(@ptrCast(self), @intCast(id));
     }
 
@@ -730,7 +730,7 @@ pub const qabstractanimation = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qobject.html#startTimer)
     ///
     /// ``` self: QtC.QAbstractAnimation, interval: i32, timerType: qnamespace_enums.TimerType ```
-    pub fn StartTimer22(self: ?*anyopaque, interval: i32, timerType: i64) i32 {
+    pub fn StartTimer22(self: ?*anyopaque, interval: i32, timerType: i32) i32 {
         return qtc.QObject_StartTimer22(@ptrCast(self), @intCast(interval), @intCast(timerType));
     }
 
@@ -739,7 +739,7 @@ pub const qabstractanimation = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qobject.html#connect)
     ///
     /// ``` sender: QtC.QObject, signal: QtC.QMetaMethod, receiver: QtC.QObject, method: QtC.QMetaMethod, typeVal: qnamespace_enums.ConnectionType ```
-    pub fn Connect5(sender: ?*anyopaque, signal: ?*anyopaque, receiver: ?*anyopaque, method: ?*anyopaque, typeVal: i64) QtC.QMetaObject__Connection {
+    pub fn Connect5(sender: ?*anyopaque, signal: ?*anyopaque, receiver: ?*anyopaque, method: ?*anyopaque, typeVal: i32) QtC.QMetaObject__Connection {
         return qtc.QObject_Connect5(@ptrCast(sender), @ptrCast(signal), @ptrCast(receiver), @ptrCast(method), @intCast(typeVal));
     }
 
@@ -748,7 +748,7 @@ pub const qabstractanimation = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qobject.html#connect)
     ///
     /// ``` self: QtC.QAbstractAnimation, sender: QtC.QObject, signal: []const u8, member: []const u8, typeVal: qnamespace_enums.ConnectionType ```
-    pub fn Connect4(self: ?*anyopaque, sender: ?*anyopaque, signal: []const u8, member: []const u8, typeVal: i64) QtC.QMetaObject__Connection {
+    pub fn Connect4(self: ?*anyopaque, sender: ?*anyopaque, signal: []const u8, member: []const u8, typeVal: i32) QtC.QMetaObject__Connection {
         const signal_Cstring = signal.ptr;
         const member_Cstring = member.ptr;
         return qtc.QObject_Connect4(@ptrCast(self), @ptrCast(sender), signal_Cstring, member_Cstring, @intCast(typeVal));
@@ -1155,21 +1155,21 @@ pub const qanimationdriver = struct {
     }
 
     /// ``` self: QtC.QAnimationDriver, param1: qobjectdefs_enums.Call, param2: i32, param3: ?*anyopaque ```
-    pub fn Metacall(self: ?*anyopaque, param1: i64, param2: i32, param3: ?*anyopaque) i32 {
+    pub fn Metacall(self: ?*anyopaque, param1: i32, param2: i32, param3: ?*anyopaque) i32 {
         return qtc.QAnimationDriver_Metacall(@ptrCast(self), @intCast(param1), @intCast(param2), param3);
     }
 
     /// Allows for overriding the related default method
     ///
     /// ``` self: QtC.QAnimationDriver, slot: fn (self: QtC.QAnimationDriver, param1: qobjectdefs_enums.Call, param2: i32, param3: ?*anyopaque) callconv(.c) i32 ```
-    pub fn OnMetacall(self: ?*anyopaque, slot: fn (?*anyopaque, i64, i32, ?*anyopaque) callconv(.c) i32) void {
+    pub fn OnMetacall(self: ?*anyopaque, slot: fn (?*anyopaque, i32, i32, ?*anyopaque) callconv(.c) i32) void {
         qtc.QAnimationDriver_OnMetacall(@ptrCast(self), @as(isize, @bitCast(@intFromPtr(&slot))));
     }
 
     /// Base class method implementation
     ///
     /// ``` self: QtC.QAnimationDriver, param1: qobjectdefs_enums.Call, param2: i32, param3: ?*anyopaque ```
-    pub fn QBaseMetacall(self: ?*anyopaque, param1: i64, param2: i32, param3: ?*anyopaque) i32 {
+    pub fn QBaseMetacall(self: ?*anyopaque, param1: i32, param2: i32, param3: ?*anyopaque) i32 {
         return qtc.QAnimationDriver_QBaseMetacall(@ptrCast(self), @intCast(param1), @intCast(param2), param3);
     }
 
@@ -1497,7 +1497,7 @@ pub const qanimationdriver = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qobject.html#killTimer)
     ///
     /// ``` self: QtC.QAnimationDriver, id: qnamespace_enums.TimerId ```
-    pub fn KillTimer2(self: ?*anyopaque, id: i64) void {
+    pub fn KillTimer2(self: ?*anyopaque, id: i32) void {
         qtc.QObject_KillTimer2(@ptrCast(self), @intCast(id));
     }
 
@@ -1720,7 +1720,7 @@ pub const qanimationdriver = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qobject.html#startTimer)
     ///
     /// ``` self: QtC.QAnimationDriver, interval: i32, timerType: qnamespace_enums.TimerType ```
-    pub fn StartTimer22(self: ?*anyopaque, interval: i32, timerType: i64) i32 {
+    pub fn StartTimer22(self: ?*anyopaque, interval: i32, timerType: i32) i32 {
         return qtc.QObject_StartTimer22(@ptrCast(self), @intCast(interval), @intCast(timerType));
     }
 
@@ -1729,7 +1729,7 @@ pub const qanimationdriver = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qobject.html#connect)
     ///
     /// ``` sender: QtC.QObject, signal: QtC.QMetaMethod, receiver: QtC.QObject, method: QtC.QMetaMethod, typeVal: qnamespace_enums.ConnectionType ```
-    pub fn Connect5(sender: ?*anyopaque, signal: ?*anyopaque, receiver: ?*anyopaque, method: ?*anyopaque, typeVal: i64) QtC.QMetaObject__Connection {
+    pub fn Connect5(sender: ?*anyopaque, signal: ?*anyopaque, receiver: ?*anyopaque, method: ?*anyopaque, typeVal: i32) QtC.QMetaObject__Connection {
         return qtc.QObject_Connect5(@ptrCast(sender), @ptrCast(signal), @ptrCast(receiver), @ptrCast(method), @intCast(typeVal));
     }
 
@@ -1738,7 +1738,7 @@ pub const qanimationdriver = struct {
     /// [Qt documentation](https://doc.qt.io/qt-6/qobject.html#connect)
     ///
     /// ``` self: QtC.QAnimationDriver, sender: QtC.QObject, signal: []const u8, member: []const u8, typeVal: qnamespace_enums.ConnectionType ```
-    pub fn Connect4(self: ?*anyopaque, sender: ?*anyopaque, signal: []const u8, member: []const u8, typeVal: i64) QtC.QMetaObject__Connection {
+    pub fn Connect4(self: ?*anyopaque, sender: ?*anyopaque, signal: []const u8, member: []const u8, typeVal: i32) QtC.QMetaObject__Connection {
         const signal_Cstring = signal.ptr;
         const member_Cstring = member.ptr;
         return qtc.QObject_Connect4(@ptrCast(self), @ptrCast(sender), signal_Cstring, member_Cstring, @intCast(typeVal));
